+++ exonum-java-binding-core/src/test/java/com/exonum/binding/storage/indices/CheckedMapProofMatcher.java
@@ -16,9 +16,9 @@
 
 package com.exonum.binding.storage.indices;
 
-import com.exonum.binding.common.proofs.map.flat.CheckedMapProof;
-import com.exonum.binding.common.proofs.map.flat.MapEntry;
-import com.exonum.binding.common.proofs.map.flat.ProofStatus;
+import com.exonum.binding.common.proofs.map.CheckedMapProof;
+import com.exonum.binding.common.proofs.map.MapEntry;
+import com.exonum.binding.common.proofs.map.ProofStatus;
 import com.exonum.binding.common.serialization.StandardSerializers;
 import com.google.common.io.BaseEncoding;
 import java.util.Arrays;